requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework>=3.12.0,<3.13.0
psycopg2>=2.8.6,<2.9.0
flake8>=3.8.0,<3.9.0
requests>=2.25.0,<2.26.0